requirements.txt
torch>=1.9.0
torchvision>=0.10.0
tensorboard>=2.5.0
PyYAML>=6.0
numpy>=1.20.0
scipy>=1.7.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
tqdm>=4.61.0
pillow>=8.2.0
einops>=0.3.2
timm>=0.4.12